+++ base_dj/models/__init__.py
@@ -1,4 +1,8 @@
 from . import base
-from . import dj
+from . import dj_template
+from . import dj_genre
+from . import dj_equalizer
+from . import dj_compilation
+from . import dj_song
 from . import company
 from . import ir_model
