+++ base_dj/models/dj_song.py
@@ -3,18 +3,10 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl)
 
 import csv
-import zipfile
-import io
-import jinja2
-import os
-import datetime
-import time
 from cStringIO import StringIO
 
 from odoo import models, fields, api, exceptions, _
 from odoo.tools.safe_eval import safe_eval, test_python_expr
-from odoo.modules.module import get_module_resource
-from odoo.addons.website.models.website import slugify
 
 
 SPECIAL_FIELDS = [
@@ -36,7 +28,6 @@ SPECIAL_FIELDS = [
 ADDONS_BLACKLIST = (
     # useless to track these modules amongst installed addons
     # TODO: anything else to ignore?
-    # XXX: shall we exclude modules that are installed via config settings?
     'base',
     'base_action_rule',
     'base_import',
@@ -135,231 +126,6 @@ DEFAULT_PYTHON_CODE = """# Available variable:
 # switch automatically to `load_csv_heavy
 # when this amount of records is reached
 # HEAVY_IMPORT_THRESHOLD = 1000
-
-
-class EqualizerXMLID(models.Model):
-    """Hold models' configuration to generate xmlid."""
-
-    _name = 'dj.equalizer.xmlid'
-
-    model = fields.Char()
-    xmlid_fields = fields.Char()
-
-    @api.multi
-    def get_xmlid_fields(self):
-        self.ensure_one()
-        return [x.strip() for x in self.xmlid_fields.split(',')
-                if x.strip()]
-
-
-class TemplateMixin(models.AbstractModel):
-    """Provide Jinja rendering capabilities."""
-
-    _name = 'dj.template.mixin'
-
-    template_path = fields.Char(
-        default=lambda self: self._default_dj_template_path,
-        required=True,
-    )
-
-    _default_dj_template_path = ''
-
-    @api.multi
-    def dj_template_vars(self):
-        """Return context variables to render template."""
-        self.ensure_one()
-        return {}
-
-    @api.multi
-    def dj_template(self, path=None):
-        """Retrieve Jinja template."""
-        self.ensure_one()
-        path = path or self.template_path
-        # load Jinja template
-        mod, filepath = path.split(':')
-        filepath = get_module_resource(mod, filepath)
-        if not filepath:
-            raise LookupError(
-                _('Template not found: `%s`') % self.template_path)
-        path, filename = os.path.split(filepath)
-        return jinja2.Environment(
-            loader=jinja2.FileSystemLoader(path)
-        ).get_template(filename)
-
-    def dj_render_template(self, template_vars=None):
-        template = self.dj_template()
-        template_vars = template_vars or self.dj_template_vars()
-        return template.render(**template_vars)
-
-
-class Genre(models.Model):
-    """Pick your favourite music genre."""
-
-    _name = 'dj.genre'
-
-    name = fields.Char(required=True, help='Name will be normalized.')
-
-    _sql_constraints = [
-        ('name_uniq', 'unique (name)', _('The name must be unique')),
-    ]
-
-    @api.model
-    def create(self, vals):
-        if vals.get('name'):
-            vals['name'] = slugify(vals['name']).replace('-', '_')
-        return super(Genre, self).create(vals)
-
-    @api.multi
-    def write(self, vals):
-        if vals.get('name'):
-            vals['name'] = slugify(vals['name']).replace('-', '_')
-        return super(Genre, self).write(vals)
-
-
-class DJcompilation(models.Model):
-    """Create compilations of songs and burn them."""
-
-    _name = 'dj.compilation'
-    _inherit = 'dj.template.mixin'
-    _default_dj_template_path = 'base_dj:discs/disc.tmpl'
-
-    name = fields.Char()
-    genre_id = fields.Many2one(
-        string='Genre',
-        comodel_name='dj.genre',
-        required=True,
-    )
-    genre = fields.Char(related='genre_id.name')
-    data_mode = fields.Selection(
-        selection=[
-            ('install', 'Install'),
-            ('demo', 'Demo'),
-        ],
-        default='install',
-    )
-    song_ids = fields.One2many('dj.song', 'compilation_id')
-    disc_path = fields.Char(
-        default='songs/{data_mode}/generated/{genre}.py',
-        required=True,
-    )
-    download_url = fields.Char(compute='_compute_download_url')
-
-    @api.multi
-    @api.depends()
-    def _compute_download_url(self):
-        for item in self:
-            item.download_url = \
-                u'/dj/download/compilation/{}'.format(item.id)
-
-    @api.multi
-    def dj_template_vars(self):
-        """Return context variables to render disc's template."""
-        self.ensure_one()
-        values = super(DJcompilation, self).dj_template_vars()
-        values.update({
-            # get all songs but scratchable ones
-            'songs': self.song_ids.filtered(lambda x: not x.scratchable())
-        })
-        return values
-
-    def _is_multicompany_env(self):
-        return bool(self.env['res.company'].search_count([]) > 1)
-
-    @api.model
-    def check_company_codename(self):
-        """Check company short codenames have been setup in multi company.
-
-        We need those to create unique codenames
-        """
-        if not self._is_multicompany_env():
-            return
-        companies = self.env['res.company'].search([('aka', '=', False)])
-        if companies:
-            raise exceptions.UserError(
-                _("Companies miss `aka` unique code: %s") % ', '.join(
-                    companies.mapped('name')
-                )
-            )
-
-    @api.multi
-    def get_all_tracks(self):
-        """Return all files to burn into the compilation."""
-        self.ensure_one()
-        files = []
-        for song in self.song_ids:
-            track = song.burn_track()
-            if track:
-                files.append(track)
-        # add __init__..py to song module folder
-        init_file = os.path.join(
-            os.path.dirname(self.disc_full_path()), '__init__.py')
-        files.append((init_file, '#'))
-        files.append(self.burn_disc())
-        files.append(self.burn_dev_readme())
-        return files
-
-    def disc_full_path(self):
-        return self.disc_path.format(**self.read()[0])
-
-    @api.multi
-    def burn_disc(self):
-        """Burn the disc with songs."""
-        self.ensure_one()
-        content = self.dj_render_template()
-        # make sure PEP8 is safe
-        # no triple empty line, only an empty line at the end
-        content = content.replace('\n\n\n\n', '\n\n\n').strip() + '\n'
-        return self.disc_full_path(), content
-
-    @api.multi
-    def burn_dev_readme(self):
-        """Burn and additional readme for developers."""
-        self.ensure_one()
-        template = self.dj_template(path='base_dj:discs/DEV_README.tmpl')
-        return 'DEV_README.rst', template.render(compilation=self)
-
-    @api.multi
-    def burn(self):
-        """Burn disc into a zip file."""
-        self.ensure_one()
-        files = self.get_all_tracks()
-        in_mem_zip = io.BytesIO()
-        with zipfile.ZipFile(in_mem_zip, "w", zipfile.ZIP_DEFLATED) as zf:
-            for filepath, data in files:
-                # File "/usr/lib/python2.7/zipfile.py", line 1247, in writestr
-                # TypeError: 'unicode' does not have the buffer interface
-                if isinstance(data, unicode):
-                    data = data.encode('utf-8')
-                # use info to keep date and set permissions
-                info = zipfile.ZipInfo(
-                    filepath, date_time=time.localtime(time.time()))
-                # set proper permissions
-                info.external_attr = 0644 << 16L
-                zf.writestr(info, data)
-        in_mem_zip.seek(0)
-        filename = self.make_album_title()
-        return filename, in_mem_zip.read()
-
-    def make_album_title(self):
-        dt = datetime.datetime.now().strftime('%Y%m%d_%H%M')
-        return '{}_{}-{}.zip'.format(
-            slugify(self.name).replace('-', '_'),
-            self.data_mode, dt)
-
-    @api.multi
-    def download_compilation(self):
-        """Download zip file."""
-        self.check_company_codename()
-        self.ensure_one()
-        return {
-            'type': 'ir.actions.act_url',
-            'target': 'new',
-            'url': self.download_url,
-        }
-
-    def anthem_path(self):
-        path = self.disc_full_path().replace('/', '.').replace('.py', '')
-        return '{}::main'.format(path)
 
 
 class Song(models.Model):
@@ -623,14 +389,14 @@ class Song(models.Model):
             model_id = vals.get('model_id') or self.model_id.id
             fields = self._get_fields(model_id, vals.pop('field_list'))
             vals['model_fields_ids'] = [(6, 0, fields.ids)]
-        return super(Song ,self).write(vals)
+        return super(Song, self).write(vals)
 
     @api.model
     def create(self, vals):
         if vals.get('field_list') and vals.get('model_id'):
             fields = self._get_fields(vals['model_id'], vals.pop('field_list'))
             vals['model_fields_ids'] = [(6, 0, fields.ids)]
-        return super(Song ,self).create(vals)
+        return super(Song, self).create(vals)
 
     def _get_fields(self, model_id, field_list):
         """ helper to set fields from a list """
@@ -639,7 +405,7 @@ class Song(models.Model):
         return self.env['ir.model.fields'].search([
             ('model_id', '=', model_name),
             ('name', 'in', field_names)
-            ])
+        ])
 
     def _get_all_fields(self):
         names = set(
